+++ packages/search/src/common/create-filters.ts
@@ -1,4 +1,5 @@
 import { ISearchParams } from "./params";
+import { paramSchema } from "./param-schema";
 
 /**
  * Create filters
@@ -42,86 +43,3 @@ function generateFilter(values: string, paramDefinition: any) {
     };
   }
 }
-
-/*
- * This object will likely live in a separate file in the near future. We may even tie into the backend
- * schema so the UI always stays in sync (perhaps with a processing step)
- * catalogDefinition: true means a field is part of defining the target set of items to search before any filtering
- */
-const paramSchema: any = {
-  q: { type: "simple" },
-  page: { type: "simple" },
-  tags: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "all"
-  },
-  source: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any"
-  },
-  bbox: { type: "simple" },
-  location_name: { type: "simple" },
-  sort: { type: "simple" },
-  groupIds: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any",
-    catalogDefinition: true
-  },
-  catalog: { type: "simple" },
-  owner: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any"
-  },
-  access: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any"
-  },
-  type: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any"
-  },
-  hubType: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any"
-  },
-  downloadable: {
-    type: "filter",
-    dataType: "boolean"
-  },
-  hasApi: {
-    type: "filter",
-    dataType: "boolean"
-  },
-  openData: {
-    type: "filter",
-    dataType: "boolean"
-  },
-  id: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any",
-    catalogDefinition: true
-  },
-  collection: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any"
-  },
-  sector: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any"
-  },
-  region: {
-    type: "filter",
-    dataType: "string",
-    defaultOp: "any"
-  }
-};
